Make seq optional in order rejected
One of the rejected reasons is service temporarily unavailable,
meaning that backend is down. When backend is down, there will be no
sequence number, so the response will be dropped, and API clients
become confused (and rightly so).

I saw a number of sentry errors about this during the other week's
backend settlement issues, and another one Saturday night.

Signed-off-by: Robin Sonefors <robin.sonefors@smarkets.com>
Reviewed-by: Daniel LR